Fixed type conversions: Removed unnecessary float to double conversion

+++ adma_ros2_driver/include/adma_ros2_driver/parser/parser_utils.hpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-const float PI=3.1415926535897932384626433832795028841971f;
+const double PI=3.1415926535897932384626433832795028841971;
 
 bool getbit(unsigned char byte, int position);
-float getScaledValue(int32_t rawValue, float lsbFactor);
+double getScaledValue(int32_t rawValue, double lsbFactor);
